backend/requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
cryptography==46.0.3
PyNaCl==1.6.1
argon2-cffi==25.1.0
python-multipart==0.0.20
aiofiles==25.1.0
Pillow>=10.0.0